refactor(dev-infra): move `parseCommitMessagesForRange` into the parse file
When attempting to actually rely on `parseCommitMessagesForRange`, it became apparent
that the function really belongs in the parse file, rather than utils.
